Code refactoring for acl controller, adding open api spec
