+++ src/main/java/io/aiven/klaw/service/AclControllerService.java
@@ -4,10 +4,10 @@ import static io.aiven.klaw.model.MailType.ACL_DELETE_REQUESTED;
 import static io.aiven.klaw.model.MailType.ACL_REQUESTED;
 import static io.aiven.klaw.model.MailType.ACL_REQUEST_APPROVED;
 import static io.aiven.klaw.model.MailType.ACL_REQUEST_DENIED;
-import static io.aiven.klaw.service.KwConstants.RETRIEVE_SCHEMAS_KEY;
 import static org.springframework.beans.BeanUtils.copyProperties;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import io.aiven.klaw.config.ManageDatabase;
@@ -15,22 +15,17 @@ import io.aiven.klaw.dao.Acl;
 import io.aiven.klaw.dao.AclRequests;
 import io.aiven.klaw.dao.Env;
 import io.aiven.klaw.dao.KwClusters;
-import io.aiven.klaw.dao.Team;
 import io.aiven.klaw.dao.Topic;
 import io.aiven.klaw.dao.UserInfo;
 import io.aiven.klaw.error.KlawException;
 import io.aiven.klaw.helpers.HandleDbRequests;
 import io.aiven.klaw.model.AclInfo;
-import io.aiven.klaw.model.AclOperationType;
 import io.aiven.klaw.model.AclRequestsModel;
 import io.aiven.klaw.model.ApiResponse;
 import io.aiven.klaw.model.ApiResultStatus;
 import io.aiven.klaw.model.KafkaClustersType;
-import io.aiven.klaw.model.KafkaFlavors;
 import io.aiven.klaw.model.PermissionType;
 import io.aiven.klaw.model.RequestStatus;
-import io.aiven.klaw.model.SyncAclUpdates;
-import io.aiven.klaw.model.SyncBackAcls;
 import io.aiven.klaw.model.TopicHistory;
 import io.aiven.klaw.model.TopicInfo;
 import io.aiven.klaw.model.TopicOverview;
@@ -163,176 +158,6 @@ public class AclControllerService {
           commonUtilsService.getLoginUrl());
     }
     return "{\"result\":\"" + execRes + "\"}";
-  }
-
-  public Map<String, String> updateSyncAcls(List<SyncAclUpdates> syncAclUpdates) {
-    log.info("updateSyncAcls {}", syncAclUpdates);
-    String userDetails = getUserName();
-    Map<String, String> response = new HashMap<>();
-    int tenantId = commonUtilsService.getTenantId(getUserName());
-
-    if (commonUtilsService.isNotAuthorizedUser(getPrincipal(), PermissionType.SYNC_SUBSCRIPTIONS)) {
-      response.put("result", "Not Authorized.");
-      return response;
-    }
-
-    List<Acl> listTopics = new ArrayList<>();
-    Acl t;
-
-    if (syncAclUpdates != null && syncAclUpdates.size() > 0) {
-
-      Map<String, SyncAclUpdates> stringSyncAclUpdatesHashMap = new HashMap<>();
-
-      // remove duplicates
-      for (SyncAclUpdates syncAclUpdateItem : syncAclUpdates) {
-        stringSyncAclUpdatesHashMap.remove(syncAclUpdateItem.getSequence());
-        stringSyncAclUpdatesHashMap.put(syncAclUpdateItem.getSequence(), syncAclUpdateItem);
-      }
-
-      for (Map.Entry<String, SyncAclUpdates> stringSyncAclUpdatesEntry :
-          stringSyncAclUpdatesHashMap.entrySet()) {
-        SyncAclUpdates syncAclUpdateItem = stringSyncAclUpdatesEntry.getValue();
-
-        // tenant filtering
-        if (!getEnvsFromUserId(userDetails).contains(syncAclUpdateItem.getEnvSelected())) {
-          response.put("result", "Not Authorized.");
-          return response;
-        }
-
-        t = new Acl();
-
-        if (syncAclUpdateItem.getReq_no() != null)
-          t.setReq_no(Integer.parseInt(syncAclUpdateItem.getReq_no()));
-        t.setTopicname(syncAclUpdateItem.getTopicName());
-        t.setConsumergroup(syncAclUpdateItem.getConsumerGroup());
-        t.setAclip(syncAclUpdateItem.getAclIp());
-        t.setAclssl(syncAclUpdateItem.getAclSsl());
-        t.setTeamId(
-            manageDatabase.getTeamIdFromTeamName(tenantId, syncAclUpdateItem.getTeamSelected()));
-        t.setEnvironment(syncAclUpdateItem.getEnvSelected());
-        t.setTopictype(syncAclUpdateItem.getAclType());
-        t.setAclPatternType("LITERAL");
-        t.setTenantId(tenantId);
-
-        listTopics.add(t);
-      }
-    } else {
-      response.put("result", "No record updated.");
-      return response;
-    }
-
-    try {
-      if (listTopics.size() > 0) {
-        response.put("result", manageDatabase.getHandleDbRequests().addToSyncacls(listTopics));
-      }
-    } catch (Exception e) {
-      log.error("Exception:", e);
-      response.put("result", "Failure." + e);
-    }
-    return response;
-  }
-
-  public Map<String, List<String>> updateSyncBackAcls(SyncBackAcls syncBackAcls) {
-    log.info("updateSyncBackAcls {}", syncBackAcls);
-    Map<String, List<String>> resultMap = new HashMap<>();
-
-    List<String> logArray = new ArrayList<>();
-    int tenantId = commonUtilsService.getTenantId(getUserName());
-
-    logArray.add(
-        "Source Environment " + getEnvDetails(syncBackAcls.getSourceEnv(), tenantId).getName());
-    logArray.add(
-        "Target Environment " + getEnvDetails(syncBackAcls.getTargetEnv(), tenantId).getName());
-    logArray.add("Type of Sync " + syncBackAcls.getTypeOfSync());
-
-    if (commonUtilsService.isNotAuthorizedUser(
-        getPrincipal(), PermissionType.SYNC_BACK_SUBSCRIPTIONS)) {
-      List<String> notAuth = new ArrayList<>();
-      notAuth.add(ApiResultStatus.NOT_AUTHORIZED.value);
-      resultMap.put("result", notAuth);
-      return resultMap;
-    }
-
-    List<String> resultStatus = new ArrayList<>();
-    resultStatus.add("success");
-
-    resultMap.put("result", resultStatus);
-
-    if ("SELECTED_ACLS".equals(syncBackAcls.getTypeOfSync())) {
-      for (String aclId : syncBackAcls.getAclIds()) {
-        Acl acl =
-            manageDatabase
-                .getHandleDbRequests()
-                .selectSyncAclsFromReqNo(Integer.parseInt(aclId), tenantId);
-        if (acl != null) {
-          approveSyncBackAcls(syncBackAcls, resultMap, logArray, acl, tenantId);
-        }
-      }
-    } else {
-      List<Acl> acls =
-          manageDatabase.getHandleDbRequests().getSyncAcls(syncBackAcls.getSourceEnv(), tenantId);
-      for (Acl acl : acls) {
-        approveSyncBackAcls(syncBackAcls, resultMap, logArray, acl, tenantId);
-      }
-    }
-
-    resultMap.put("syncbacklog", logArray);
-
-    return resultMap;
-  }
-
-  private void approveSyncBackAcls(
-      SyncBackAcls syncBackAcls,
-      Map<String, List<String>> resultMap,
-      List<String> logUpdateSyncBackTopics,
-      Acl aclFound,
-      int tenantId) {
-    try {
-      AclRequests aclReq = new AclRequests();
-      copyProperties(aclFound, aclReq);
-      aclReq.setReq_no(null);
-      aclReq.setAcl_ip(aclFound.getAclip());
-      aclReq.setAcl_ssl(aclFound.getAclssl());
-      aclReq.setEnvironment(syncBackAcls.getTargetEnv());
-      aclReq.setRequestingteam(aclFound.getTeamId());
-      aclReq.setAclType(AclOperationType.CREATE.value);
-      aclReq.setUsername(getUserName());
-      aclReq.setTenantId(tenantId);
-
-      ResponseEntity<ApiResponse> response = clusterApiService.approveAclRequests(aclReq, tenantId);
-
-      ApiResponse responseBody = response.getBody();
-      //      String resultAclReq = responseBody.getResult();
-      String resultAclNullCheck = Objects.requireNonNull(responseBody).getResult();
-      if (!Objects.requireNonNull(resultAclNullCheck).contains("success")) {
-        log.error("Error in creating acl {} {}", aclFound, responseBody);
-        logUpdateSyncBackTopics.add(
-            "Error in Acl creation. Acl:" + aclFound.getTopicname() + " " + resultAclNullCheck);
-      } else if (resultAclNullCheck.contains("Acl already exists")) {
-        logUpdateSyncBackTopics.add("Acl already exists " + aclFound.getTopicname());
-      } else {
-        if (!Objects.equals(syncBackAcls.getSourceEnv(), syncBackAcls.getTargetEnv())) {
-          logUpdateSyncBackTopics.add("Acl added: " + aclFound.getTopicname());
-          // Create request
-          Map<String, String> resultMapReq =
-              manageDatabase.getHandleDbRequests().requestForAcl(aclReq);
-          if (resultMapReq.containsKey("aclId")) {
-            Integer aclId = Integer.parseInt(resultMapReq.get("aclId"));
-            aclReq.setReq_no(aclId);
-            // Approve request
-            String emptyJsonParams = "{}";
-            manageDatabase
-                .getHandleDbRequests()
-                .updateAclRequest(aclReq, getUserName(), emptyJsonParams);
-          }
-        }
-      }
-    } catch (KlawException e) {
-      log.error("Error in creating acl {} {}", e.getMessage(), aclFound);
-      List<String> resultStatus = new ArrayList<>();
-      resultStatus.add("Error :" + e.getMessage());
-      resultMap.put("result", resultStatus);
-    }
   }
 
   public List<AclRequestsModel> getAclRequests(
@@ -746,60 +571,6 @@ public class AclControllerService {
     } else return "{\"result\":\"Record not found !\"}";
   }
 
-  private List<Map<String, String>> getAclListFromCluster(
-      String bootstrapHost,
-      Env envSelected,
-      String protocol,
-      String clusterName,
-      String topicNameSearch,
-      int tenantId)
-      throws KlawException {
-    List<Map<String, String>> aclList;
-    aclList =
-        clusterApiService.getAcls(bootstrapHost, envSelected, protocol, clusterName, tenantId);
-    return updateConsumerGroups(groupAcls(aclList, topicNameSearch, true), aclList);
-  }
-
-  private List<Map<String, String>> updateConsumerGroups(
-      List<Map<String, String>> groupedList, List<Map<String, String>> clusterAclList) {
-    List<Map<String, String>> updateList = new ArrayList<>(groupedList);
-
-    for (Map<String, String> hMapGroupItem : groupedList) {
-      for (Map<String, String> hMapItem : clusterAclList) {
-        if ("READ".equals(hMapGroupItem.get("operation"))
-            && "READ".equals(hMapItem.get("operation"))
-            && "GROUP".equals(hMapItem.get("resourceType"))) {
-          if (Objects.equals(hMapItem.get("host"), hMapGroupItem.get("host"))
-              && Objects.equals(hMapItem.get("principle"), hMapGroupItem.get("principle"))) {
-            Map<String, String> hashMap = new HashMap<>(hMapGroupItem);
-            hashMap.put("consumerGroup", hMapItem.get("resourceName"));
-            updateList.add(hashMap);
-            break;
-          }
-        }
-      }
-    }
-    return updateList;
-  }
-
-  private List<Map<String, String>> groupAcls(
-      List<Map<String, String>> aclList, String topicNameSearch, boolean isSync) {
-
-    return aclList.stream()
-        .filter(
-            hItem -> {
-              if (isSync) {
-                if (topicNameSearch != null) {
-                  return "TOPIC".equals(hItem.get("resourceType"))
-                      && hItem.get("resourceName").contains(topicNameSearch);
-                } else return "TOPIC".equals(hItem.get("resourceType"));
-              } else {
-                return Objects.equals(hItem.get("resourceName"), topicNameSearch);
-              }
-            })
-        .collect(Collectors.toList());
-  }
-
   private List<Acl> getAclsFromSOT(
       String env, String topicNameSearch, boolean regex, int tenantId) {
     List<Acl> aclsFromSOT;
@@ -888,8 +659,6 @@ public class AclControllerService {
     String userDetails = getUserName();
     HandleDbRequests handleDb = manageDatabase.getHandleDbRequests();
     int tenantId = commonUtilsService.getTenantId(getUserName());
-    String retrieveSchemasStr = manageDatabase.getKwPropertyValue(RETRIEVE_SCHEMAS_KEY, tenantId);
-    boolean retrieveSchemas = "true".equals(retrieveSchemasStr);
 
     if (topicNameSearch != null) topicNameSearch = topicNameSearch.trim();
     else return null;
@@ -948,7 +717,8 @@ public class AclControllerService {
 
       if (topic.getHistory() != null) {
         try {
-          topicHistoryFromTopic = OBJECT_MAPPER.readValue(topic.getHistory(), ArrayList.class);
+          topicHistoryFromTopic =
+              OBJECT_MAPPER.readValue(topic.getHistory(), new TypeReference<>() {});
           topicHistoryList.addAll(topicHistoryFromTopic);
         } catch (JsonProcessingException e) {
           log.error("Unable to parse topicHistory");
@@ -1128,7 +898,8 @@ public class AclControllerService {
               schemaMap.put("nextVersion", "" + allVersionsList.get(indexOfVersion + 1));
             }
           } else {
-            hashMapSchemaObj = schemaObjects.get(Integer.parseInt(schemaVersionSearch));
+            hashMapSchemaObj =
+                schemaObjects.get(Integer.parseInt(Objects.requireNonNull(schemaVersionSearch)));
             schemaOfObj = (String) hashMapSchemaObj.get("schema");
             schemaMap.put("isLatest", "false");
             schemaMap.put("id", hashMapSchemaObj.get("id") + "");
@@ -1166,74 +937,6 @@ public class AclControllerService {
     }
   }
 
-  public List<AclInfo> getSyncAcls(
-      String env, String pageNo, String currentPage, String topicNameSearch, String showAllAcls)
-      throws KlawException {
-    log.info(
-        "getSyncAcls env: {} topicNameSearch: {} showAllAcls:{}",
-        env,
-        topicNameSearch,
-        showAllAcls);
-    boolean isReconciliation = !Boolean.parseBoolean(showAllAcls);
-    int tenantId = commonUtilsService.getTenantId(getUserName());
-
-    if (topicNameSearch != null) topicNameSearch = topicNameSearch.trim();
-
-    if (commonUtilsService.isNotAuthorizedUser(getPrincipal(), PermissionType.SYNC_SUBSCRIPTIONS))
-      return null;
-
-    List<Map<String, String>> aclList;
-
-    Env envSelected = getEnvDetails(env, tenantId);
-    KwClusters kwClusters =
-        manageDatabase
-            .getClusters(KafkaClustersType.KAFKA.value, tenantId)
-            .get(envSelected.getClusterId());
-    aclList =
-        getAclListFromCluster(
-            kwClusters.getBootstrapServers(),
-            envSelected,
-            kwClusters.getProtocol(),
-            kwClusters.getClusterName(),
-            topicNameSearch,
-            tenantId);
-
-    List<Acl> aclsFromSOT = getAclsFromSOT(env, topicNameSearch, true, tenantId);
-
-    topicCounter = 0;
-    return getAclsList(
-        pageNo,
-        currentPage,
-        applyFiltersAcls(
-            env, aclList, aclsFromSOT, isReconciliation, tenantId, kwClusters.getKafkaFlavor()));
-  }
-
-  public List<AclInfo> getSyncBackAcls(
-      String envId, String pageNo, String currentPage, String topicNameSearch, String teamName) {
-    log.info("getSyncBackAcls {} {} {} {}", envId, pageNo, topicNameSearch, teamName);
-    if (topicNameSearch != null) topicNameSearch = topicNameSearch.trim();
-
-    int tenantId = commonUtilsService.getTenantId(getUserName());
-
-    if (commonUtilsService.isNotAuthorizedUser(
-        getPrincipal(), PermissionType.SYNC_BACK_SUBSCRIPTIONS)) return null;
-
-    List<Acl> aclsFromSOT;
-
-    if (topicNameSearch != null && topicNameSearch.trim().length() > 0)
-      aclsFromSOT = getAclsFromSOT(envId, topicNameSearch, false, tenantId);
-    else aclsFromSOT = getAclsFromSOT(envId, topicNameSearch, true, tenantId);
-
-    List<AclInfo> aclInfoList;
-    Integer loggedInUserTeam = getMyTeamId(getUserName());
-
-    aclInfoList =
-        getAclsList(
-            pageNo, currentPage, applyFiltersAclsForSOT(loggedInUserTeam, aclsFromSOT, tenantId));
-
-    return aclInfoList;
-  }
-
   private List<AclInfo> applyFiltersAclsForSOT(
       Integer loggedInUserTeam, List<Acl> aclsFromSOT, int tenantId) {
 
@@ -1259,145 +962,6 @@ public class AclControllerService {
       if (aclSotItem.getAclip() != null || aclSotItem.getAclssl() != null) aclList.add(mp);
     }
     return aclList;
-  }
-
-  private List<AclInfo> applyFiltersAcls(
-      String env,
-      List<Map<String, String>> aclList,
-      List<Acl> aclsFromSOT,
-      boolean isReconciliation,
-      int tenantId,
-      String kafkaFlavor) {
-
-    List<AclInfo> aclListMap = new ArrayList<>();
-    List<String> teamList = new ArrayList<>();
-    teamList = tenantFiltering(teamList);
-
-    for (Map<String, String> aclListItem : aclList) {
-      AclInfo mp = new AclInfo();
-      mp.setEnvironment(env);
-
-      mp.setPossibleTeams(teamList);
-      mp.setTeamname("");
-
-      String tmpPermType = aclListItem.get("operation");
-
-      if ("WRITE".equals(tmpPermType)) mp.setTopictype("Producer");
-      else if ("READ".equals(tmpPermType)) {
-        mp.setTopictype("Consumer");
-        if (aclListItem.get("consumerGroup") != null)
-          mp.setConsumergroup(aclListItem.get("consumerGroup"));
-        else continue;
-      }
-
-      if ("topic".equalsIgnoreCase(aclListItem.get("resourceType")))
-        mp.setTopicname(aclListItem.get("resourceName"));
-
-      mp.setAcl_ip(aclListItem.get("host"));
-      mp.setAcl_ssl(aclListItem.get("principle"));
-
-      for (Acl aclSotItem : aclsFromSOT) {
-        String acl_ssl = aclSotItem.getAclssl();
-        String acl_host = aclSotItem.getAclip();
-
-        if (acl_ssl == null || acl_ssl.equals("")) acl_ssl = "User:*";
-        else {
-          if (!KafkaFlavors.AIVEN_FOR_APACHE_KAFKA.value.equals(kafkaFlavor)
-              && !"User:*".equals(acl_ssl)
-              && !acl_ssl.startsWith("User:")) {
-            acl_ssl = "User:" + acl_ssl;
-          }
-        }
-
-        if (acl_host == null || acl_host.equals("")) acl_host = "*";
-
-        if (aclSotItem.getTopicname() != null
-            && Objects.equals(aclListItem.get("resourceName"), aclSotItem.getTopicname())
-            && Objects.equals(aclListItem.get("host"), acl_host)
-            && Objects.equals(aclListItem.get("principle"), acl_ssl)
-            && Objects.equals(aclSotItem.getTopictype(), mp.getTopictype())) {
-          mp.setTeamname(manageDatabase.getTeamNameFromTeamId(tenantId, aclSotItem.getTeamId()));
-          mp.setReq_no(aclSotItem.getReq_no() + "");
-          break;
-        }
-      }
-
-      if (mp.getTeamname() == null) mp.setTeamname("Unknown");
-
-      if (isReconciliation) {
-        if ("Unknown".equals(mp.getTeamname()) || "".equals(mp.getTeamname())) aclListMap.add(mp);
-      } else {
-        if (teamList.contains(mp.getTeamname())) aclListMap.add(mp);
-        else if ("Unknown".equals(mp.getTeamname()) || "".equals(mp.getTeamname()))
-          aclListMap.add(mp);
-      }
-    }
-    return aclListMap;
-  }
-
-  private List<String> tenantFiltering(List<String> teamList) {
-    if (!commonUtilsService.isNotAuthorizedUser(
-            getPrincipal(), PermissionType.SYNC_BACK_SUBSCRIPTIONS)
-        || !commonUtilsService.isNotAuthorizedUser(getPrincipal(), PermissionType.SYNC_TOPICS)
-        || !commonUtilsService.isNotAuthorizedUser(
-            getPrincipal(), PermissionType.SYNC_SUBSCRIPTIONS)
-        || !commonUtilsService.isNotAuthorizedUser(
-            getPrincipal(), PermissionType.SYNC_BACK_TOPICS)) {
-      // tenant filtering
-      int tenantId = commonUtilsService.getTenantId(getUserName());
-
-      List<Team> teams = manageDatabase.getHandleDbRequests().selectAllTeams(tenantId);
-
-      teams =
-          teams.stream()
-              .filter(t -> Objects.equals(t.getTenantId(), tenantId))
-              .collect(Collectors.toList());
-      List<String> teamListUpdated = new ArrayList<>();
-      for (Team teamsItem : teams) {
-        teamListUpdated.add(teamsItem.getTeamname());
-      }
-      teamList = teamListUpdated;
-    }
-    return teamList;
-  }
-
-  private List<AclInfo> getAclsList(String pageNo, String currentPage, List<AclInfo> aclListMap) {
-    List<AclInfo> aclListMapUpdated = new ArrayList<>();
-
-    int totalRecs = aclListMap.size();
-    int recsPerPage = 20;
-    int totalPages =
-        aclListMap.size() / recsPerPage + (aclListMap.size() % recsPerPage > 0 ? 1 : 0);
-
-    pageNo = commonUtilsService.deriveCurrentPage(pageNo, currentPage, totalPages);
-    int requestPageNo = Integer.parseInt(pageNo);
-    int startVar = (requestPageNo - 1) * recsPerPage;
-    int lastVar = (requestPageNo) * (recsPerPage);
-
-    List<String> numList = new ArrayList<>();
-    commonUtilsService.getAllPagesList(pageNo, currentPage, totalPages, numList);
-
-    for (int i = 0; i < totalRecs; i++) {
-      int counterInc = counterIncrement();
-      if (i >= startVar && i < lastVar) {
-        AclInfo mp = aclListMap.get(i);
-        mp.setSequence(counterInc + "");
-
-        mp.setTotalNoPages(totalPages + "");
-        mp.setCurrentPage(pageNo);
-        mp.setAllPageNos(numList);
-
-        aclListMapUpdated.add(mp);
-      }
-    }
-    return aclListMapUpdated;
-  }
-
-  private int topicCounter = 0;
-
-  private int counterIncrement() {
-    topicCounter++;
-    return topicCounter;
   }
 
   private String getUserName() {
