+++ src/test/java/io/aiven/klaw/controller/AclControllerTest.java
@@ -15,6 +15,7 @@ import io.aiven.klaw.model.ApiResponse;
 import io.aiven.klaw.model.SyncAclUpdates;
 import io.aiven.klaw.model.TopicOverview;
 import io.aiven.klaw.service.AclControllerService;
+import io.aiven.klaw.service.AclSyncControllerService;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -35,11 +36,17 @@ public class AclControllerTest {
   public static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
   @MockBean private AclControllerService aclControllerService;
 
+  @MockBean private AclSyncControllerService aclSyncControllerService;
+
   private UtilMethods utilMethods;
 
-  private MockMvc mvc;
+  private MockMvc mvcAcls;
 
   private AclController aclController;
+
+  private MockMvc mvcAclsSync;
+
+  private AclSyncController aclSyncController;
 
   private static final String topicName = "testtopic";
   private static final int topicId = 1001;
@@ -47,9 +54,13 @@ public class AclControllerTest {
   @BeforeEach
   public void setup() {
     aclController = new AclController();
+    aclSyncController = new AclSyncController();
     utilMethods = new UtilMethods();
-    mvc = MockMvcBuilders.standaloneSetup(aclController).dispatchOptions(true).build();
+    mvcAcls = MockMvcBuilders.standaloneSetup(aclController).dispatchOptions(true).build();
     ReflectionTestUtils.setField(aclController, "aclControllerService", aclControllerService);
+    mvcAclsSync = MockMvcBuilders.standaloneSetup(aclSyncController).dispatchOptions(true).build();
+    ReflectionTestUtils.setField(
+        aclSyncController, "aclSyncControllerService", aclSyncControllerService);
   }
 
   @Test
@@ -60,7 +71,8 @@ public class AclControllerTest {
     when(aclControllerService.createAcl(any())).thenReturn("success");
 
     String response =
-        mvc.perform(
+        mvcAcls
+            .perform(
                 MockMvcRequestBuilders.post("/createAcl")
                     .content(jsonReq)
                     .contentType(MediaType.APPLICATION_JSON)
@@ -80,10 +92,11 @@ public class AclControllerTest {
     String jsonReq = OBJECT_MAPPER.writer().writeValueAsString(syncUpdates);
     Map<String, String> result = new HashMap<>();
     result.put("result", "success");
-    when(aclControllerService.updateSyncAcls(any())).thenReturn(result);
+    when(aclSyncControllerService.updateSyncAcls(any())).thenReturn(result);
 
     String response =
-        mvc.perform(
+        mvcAclsSync
+            .perform(
                 MockMvcRequestBuilders.post("/updateSyncAcls")
                     .content(jsonReq)
                     .contentType(MediaType.APPLICATION_JSON)
@@ -94,7 +107,7 @@ public class AclControllerTest {
             .getContentAsString();
 
     Map<String, String> actualResult =
-        new ObjectMapper().readValue(response, new TypeReference<Map<String, String>>() {});
+        new ObjectMapper().readValue(response, new TypeReference<>() {});
 
     assertThat(actualResult).containsEntry("result", "success");
   }
@@ -107,7 +120,8 @@ public class AclControllerTest {
     when(aclControllerService.getAclRequests("1", "", "all")).thenReturn(aclRequests);
 
     String res =
-        mvc.perform(
+        mvcAcls
+            .perform(
                 MockMvcRequestBuilders.get("/getAclRequests")
                     .contentType(MediaType.APPLICATION_JSON)
                     .param("pageNo", "1")
@@ -129,7 +143,8 @@ public class AclControllerTest {
     when(aclControllerService.getCreatedAclRequests("1", "", "created")).thenReturn(aclRequests);
 
     String res =
-        mvc.perform(
+        mvcAcls
+            .perform(
                 MockMvcRequestBuilders.get("/getCreatedAclRequests")
                     .param("pageNo", "1")
                     .contentType(MediaType.APPLICATION_JSON)
@@ -148,7 +163,8 @@ public class AclControllerTest {
     when(aclControllerService.deleteAclRequests(anyString())).thenReturn("success");
 
     String response =
-        mvc.perform(
+        mvcAcls
+            .perform(
                 MockMvcRequestBuilders.post("/deleteAclRequests")
                     .param("req_no", "fsda32FSDw")
                     .contentType(MediaType.APPLICATION_JSON)
@@ -167,7 +183,8 @@ public class AclControllerTest {
     when(aclControllerService.approveAclRequests(anyString())).thenReturn(apiResponse);
 
     String response =
-        mvc.perform(
+        mvcAcls
+            .perform(
                 MockMvcRequestBuilders.post("/execAclRequest")
                     .param("req_no", "reqno")
                     .contentType(MediaType.APPLICATION_JSON)
@@ -185,7 +202,8 @@ public class AclControllerTest {
     when(aclControllerService.declineAclRequests(anyString(), anyString())).thenReturn("success");
 
     String response =
-        mvc.perform(
+        mvcAcls
+            .perform(
                 MockMvcRequestBuilders.post("/execAclRequestDecline")
                     .param("req_no", "reqno")
                     .param("reasonForDecline", "reason")
@@ -206,7 +224,8 @@ public class AclControllerTest {
     when(aclControllerService.getAcls("testtopic")).thenReturn(topicOverview);
 
     String res =
-        mvc.perform(
+        mvcAcls
+            .perform(
                 MockMvcRequestBuilders.get("/getAcls")
                     .param("topicnamesearch", "testtopic")
                     .contentType(MediaType.APPLICATION_JSON)
@@ -227,7 +246,8 @@ public class AclControllerTest {
     when(aclControllerService.getAcls(null)).thenReturn(topicOverview);
 
     String res =
-        mvc.perform(
+        mvcAcls
+            .perform(
                 MockMvcRequestBuilders.get("/getAcls")
                     .contentType(MediaType.APPLICATION_JSON)
                     .accept(MediaType.APPLICATION_JSON))
@@ -241,11 +261,12 @@ public class AclControllerTest {
   public void getSyncAcls() throws Exception {
     List<AclInfo> aclInfo = utilMethods.getAclInfoList();
 
-    when(aclControllerService.getSyncAcls(anyString(), anyString(), anyString(), any(), any()))
+    when(aclSyncControllerService.getSyncAcls(anyString(), anyString(), anyString(), any(), any()))
         .thenReturn(aclInfo);
 
     String res =
-        mvc.perform(
+        mvcAclsSync
+            .perform(
                 MockMvcRequestBuilders.get("/getSyncAcls")
                     .param("env", "DEV")
                     .param("pageNo", "1")
